FEAT(<ice_strlen.c>): upgarde <ice_strlen.c>, can now return size for unsigned long long and prevent from null array.

+++ include/ice/string.h
@@ -13,7 +13,7 @@
  * @param str that will be analysed.
  * @return The len of the str.
  */
-int ice_strlen(const char *str);
+unsigned long long ice_strlen(const char *str);
 
 /**
  * @brief Compare two str.
